+++ midas/__init__.py
@@ -81,4 +81,4 @@ __all__ = [
     "IVFIndex",
     "IVFStore",
 ]
-__version__ = "0.0.2"
+__version__ = "0.0.3"